Add configuration options to determine the header values to use for platform access.
Some Pantheon customers will be managing their own proxy and will not be able to use the 'adv-cdn-origin' header
(for example, IIS is unable to add request header). For situations like this, using a different header, such as
the Host header will make more sense for determining whether the request is on/off platform.

The two configuration keys 'header' and 'matchvalue' are defaulted to 'adv-cdn-origin' and '1' to maintain downward
compatability with existing implementations using this module without those configurations.